+++ src/auth_mail/generate_token.py
