+++ src/iiif/image_server.py
@@ -6,8 +6,8 @@ from django.conf import settings
 from django.http import HttpResponse
 from requests.exceptions import RequestException
 
-from iiif import zip_tools
-from iiif.utils import ImmediateHttpResponse
+from main.utils import ImmediateHttpResponse
+from zip_consumer import zip_tools
 
 log = logging.getLogger(__name__)
