+++ src/iiif/metadata.py
@@ -5,7 +5,7 @@ from django.conf import settings
 from django.http import HttpResponse
 from requests.exceptions import RequestException
 
-from iiif.utils import ImmediateHttpResponse
+from main.utils import ImmediateHttpResponse
 
 log = logging.getLogger(__name__)
