+++ src/zip_consumer/zip_tools.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from uuid import uuid4
 from zipfile import ZipFile
 
-from iiif.utils_azure import get_queue_client
+from main.utils_azure import get_queue_client
 
 TMP_BOUWDOSSIER_ZIP_FOLDER = "/tmp/bouwdossier-zips/"
 ZIP_MESSAGE_VERSION_NAME = "zip_job_v1"
