+++ tests/test_utils_azure.py
@@ -5,7 +5,7 @@ import requests
 from azure.core.exceptions import ResourceExistsError
 from django.conf import settings
 
-from iiif.utils_azure import get_blob_service_client, get_queue_client
+from main.utils_azure import get_blob_service_client, get_queue_client
 
 
 def create_blob_container(container_name):
