+++ tests/test_zip_endpoint.py
@@ -8,16 +8,14 @@ import pytz
 from azure.core.exceptions import ResourceNotFoundError
 from django.conf import settings
 
-from iiif.authentication import (
+from auth_mail.authentication import (
     RESPONSE_CONTENT_NO_WABO_WITH_MAIL_LOGIN,
     RESPONSE_CONTENT_RESTRICTED,
     RESPONSE_CONTENT_RESTRICTED_IN_ZIP,
     create_mail_login_token,
 )
-from iiif.generate_token import create_authz_token
-from iiif.queue_zip_consumer import AzureZipQueueConsumer
-from iiif.utils_azure import get_blob_from_storage_account, get_queue_client
-from iiif.zip_tools import TMP_BOUWDOSSIER_ZIP_FOLDER
+from auth_mail.generate_token import create_authz_token
+from main.utils_azure import get_blob_from_storage_account, get_queue_client
 from tests.test_iiif import (
     IMAGE_BINARY_DATA,
     PRE_WABO_IMG_URL_BASE,
@@ -27,6 +25,8 @@ from tests.test_iiif import (
 )
 from tests.test_utils_azure import create_blob_container, create_queue
 from tests.tools import MockResponse
+from zip_consumer.queue_zip_consumer import AzureZipQueueConsumer
+from zip_consumer.zip_tools import TMP_BOUWDOSSIER_ZIP_FOLDER
 
 log = logging.getLogger(__name__)
 timezone = pytz.timezone("UTC")
@@ -353,11 +353,11 @@ class TestZipEndpoint:
         assert response.status_code == 200
         assert len(self.get_all_queue_messages()) == 1
 
-    @patch("iiif.queue_zip_consumer.create_storage_account_temp_url")
+    @patch("zip_consumer.queue_zip_consumer.create_storage_account_temp_url")
     @patch("iiif.image_server.get_image_from_server")
     @patch("iiif.metadata.do_metadata_request")
-    @patch("iiif.mailing.send_email")
-    @patch("iiif.zip_tools.cleanup_local_files")
+    @patch("auth_mail.mailing.send_email")
+    @patch("zip_consumer.zip_tools.cleanup_local_files")
     @pytest.mark.parametrize(
         "scope, second_image_access, expected_line_end, expected_files",
         [
